requirements.txt
asgiref==3.4.1
Django==3.2.8
django-admin-tools==0.9.1
django-restframework==0.0.1
djangorestframework==3.12.4
numpy==1.21.3
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
